made breakpoint changer self contained
